Checkpoints and Killzones
Checkpoint prefab added. The check points can be placed in the level. When the player enter the trigger, the checkpoint stores the players position and makes that checkpoint the new checkpoint in the level manager. Checkpoints can only be activated once. Checkpoints also store players health and reset

Killzone resets the player to the last checkpoint.